language page binding issue fixed
