+++ raaghu-mfe/projects/language/src/app/app.component.ts
@@ -1,10 +1,9 @@
 import { Component, OnInit, Input } from '@angular/core';
 import { DatePipe } from '@angular/common';
 import { AlertService, ComponentLoaderOptions } from '@libs/shared';
-import { finalize } from 'rxjs/operators';
 import { Store } from '@ngrx/store';
-import { deleteLanguage, getLanguages, setDefaultLanguage, } from 'projects/libs/state-management/src/lib/state/language/language.actions';
-import { selectAllLanguages, selectDefaultLanguage } from 'projects/libs/state-management/src/lib/state/language/language.selector';
+import { deleteLanguage, getCountryList, getLanguages, saveLanguage, setDefaultLanguage, } from '@libs/state-management';
+import { selectAllCountries, selectAllLanguages, selectDefaultLanguage } from '@libs/state-management';
 import { TableHeader } from 'projects/rds-components/src/models/table-header.model';
 import { Router } from '@angular/router';
 import { TranslateService } from '@ngx-translate/core';
@@ -15,6 +14,7 @@ import {
   style,
   animate,
 } from '@angular/animations';
+import { Item } from 'projects/libs/state-management/src/lib/state/language/language.models';
 declare var bootstrap: any;
 
 
@@ -60,7 +60,7 @@ export class AppComponent implements OnInit {
   @Input() listItems = [
     { value: 'New Language', some: 'value', key: 'new', icon: 'plus', iconWidth: '20px', iconHeight: '20px' },
   ];
-  isShimmer:boolean= false;
+  isShimmer: boolean = false;
   EditShimmer: boolean = false;
   languageCanvasTitle = 'New Language';
   public rdsAlertMfeConfig: ComponentLoaderOptions = {
@@ -89,16 +89,14 @@ export class AppComponent implements OnInit {
   ]
   loading: boolean = true;
   languageTableData: any = [];
+  public rdsNewLanguageMfeConfig: ComponentLoaderOptions;
+  public languageNames: Item[] = [];
+  public flags: any = [];
   public viewCanvas: boolean = false;
   constructor(public datepipe: DatePipe, public translate: TranslateService, private store: Store, private alertService: AlertService, private router: Router) { }
   ngOnInit(): void {
     this.isAnimation = true;
-    this.store.select(selectDefaultLanguage).subscribe((res: any) => {
-      if (res) {
-        this.translate.use(res);
-      }
-    })
-    this.subscribeToAlerts();
+
     this.rdsLanguageTableMfeConfig = {
       name: 'RdsDataTable',
       input: {
@@ -111,7 +109,7 @@ export class AppComponent implements OnInit {
         inlineEdit: false,
         actions: [{ id: 'edit', displayName: 'Edit' }, { id: 'changeText', displayName: 'Change Texts' }, { id: 'setDefaultLanguage', displayName: 'Set as default language' }, { id: 'delete', displayName: 'Delete' }],
         noDataTitle: 'Currently you do not have language',
-        isShimmer:true,
+        isShimmer: true,
       },
       output: {
         onActionSelection: (event: any) => {
@@ -119,9 +117,17 @@ export class AppComponent implements OnInit {
             this.store.dispatch(deleteLanguage(event.selectedData.id))
           } else if (event.actionId === 'edit') {
             this.languageCanvasTitle = 'Edit Language';
-            this.selectedLanguage = event.selectedData.name;
-            this.openCanvas(true);
-
+            this.selectedLanguage = JSON.parse(JSON.stringify(event.selectedData));
+            if (this.selectedLanguage.icon && this.flags.length > 0) {
+              const selectedLanguage = this.flags.find((x: any) => x.value === this.selectedLanguage.icon);
+              if (selectedLanguage) {
+                this.selectedLanguage.icon = selectedLanguage.some;
+              }
+            }
+            if (this.selectedLanguage.countryCode) {
+              this.selectedLanguage.countryCode = [this.selectedLanguage.countryCode];
+            }
+            this.openCanvas(true)
           }
           else if (event.actionId === 'setDefaultLanguage') {
             const data: any = { name: event.selectedData.countryCode };
@@ -133,11 +139,57 @@ export class AppComponent implements OnInit {
         },
       }
     };
-   this.store.dispatch(getLanguages());
+    this.rdsNewLanguageMfeConfig = {
+      name: 'RdsCompNewLanguage',
+      input: {
+        flags: this.flags,
+        languageNames: this.languageNames,
+        selectedLanguage: this.selectedLanguage,
+        EditShimmer: true,
+      },
+      output: {
+        onLanguageSave: (data: any) => {
+          const body: any = {
+            language: data
+          }
+          this.store.dispatch(saveLanguage(body));
+          this.closeCanvas();
+          // var offcanvas = document.getElementById('AddLanguage');
+          // let openedCanvas = bootstrap.Offcanvas.getInstance(offcanvas);
+          // openedCanvas.hide();
+        },
+        onCloseCanvas: (event: any) => {
+          this.closeCanvas();
+        }
+      }
+    };
+    this.store.dispatch(getCountryList());
+    this.store.dispatch(getLanguages());
+    this.store.select(selectAllCountries).subscribe((res: any) => {
+      if (res) {
+        this.languageNames = res.languageNames;
+        this.flags = [];
+        res.flags.forEach((res: any) => {
+          const icon = res.displayText.replace('-', '_')
+          this.flags.push({ some: res.displayText, value: res.value, icon: icon, iconWidth: '18px', iconHeight: '18px' })
+        });
 
+        const rdsNewLanguageMfeConfig = this.rdsNewLanguageMfeConfig;
+        rdsNewLanguageMfeConfig.input.flags = [...this.flags];
+        rdsNewLanguageMfeConfig.input.selectedLanguage = this.selectedLanguage;
+        rdsNewLanguageMfeConfig.input.languageNames = [...this.languageNames]
+        rdsNewLanguageMfeConfig.input.EditShimmer = false;
+        this.rdsNewLanguageMfeConfig = rdsNewLanguageMfeConfig;
+      }
+    });
+    this.store.select(selectDefaultLanguage).subscribe((res: any) => {
+      if (res) {
+        this.translate.use(res);
+      }
+    });
     this.store.select(selectAllLanguages).subscribe((res: any) => {
       this.languageTableData = [];
-      if (res  && res.items && res.items.length > 0 ) {
+      if (res && res.items && res.items.length > 0) {
         this.isAnimation = false;
         let defaultLanguage = res.defaultLanguageName;
         res.items.forEach((element: any) => {
@@ -153,18 +205,22 @@ export class AppComponent implements OnInit {
             statusTemplate: status,
             icon: element.icon,
             creationTime: this.datepipe.transform(new Date(element.creationTime), 'MM/dd/yyyy, h:mm:ss a'),
-            name:element.displayName
+            name: element.displayName
           }
           this.languageTableData.push(item);
         });
         const mfeConfig = this.rdsLanguageTableMfeConfig
         mfeConfig.input.tableData = [... this.languageTableData];
-        mfeConfig.input.isShimmer=false;
+        mfeConfig.input.isShimmer = false;
         this.rdsLanguageTableMfeConfig = mfeConfig;
 
 
       }
-    })
+    });
+    this.subscribeToAlerts();
+
+
+
 
 
   }
@@ -193,14 +249,13 @@ export class AppComponent implements OnInit {
     this.rdsLanguageTableMfeConfig = mfeConfig;
 
   }
-  addLanguages(language: any) {
-    this.languageTableData.push(language.NewLanguage)
-  }
+  // addLanguages(language: any) {
+  //   this.languageTableData.push(language.NewLanguage)
+  // }
 
-  openCanvas(isEdit: boolean): void {
+  openCanvas(edit: boolean = false): void {
     this.viewCanvas = true;
-
-    if (!isEdit) {
+    if (!edit) {
       this.selectedLanguage = {
         countryCode: '',
         icon: '',
@@ -208,15 +263,19 @@ export class AppComponent implements OnInit {
         id: undefined,
       };
       this.languageCanvasTitle = 'New Language';
-    }else{
+
+    } else {
       this.languageCanvasTitle = 'Edit Language';
 
     }
-    // setTimeout(() => {
-    var offcanvas = document.getElementById('AddLanguage');
-    var bsOffcanvas = new bootstrap.Offcanvas(offcanvas);
-    bsOffcanvas.show();
-    // }, 5000);
+    const rdsNewLanguageMfeConfig = this.rdsNewLanguageMfeConfig;
+    rdsNewLanguageMfeConfig.input.selectedLanguage = this.selectedLanguage;
+    this.rdsNewLanguageMfeConfig = rdsNewLanguageMfeConfig;
+    setTimeout(() => {
+      var offcanvas = document.getElementById('AddLanguage');
+      var bsOffcanvas = new bootstrap.Offcanvas(offcanvas);
+      bsOffcanvas.show();
+    }, 100);
 
   }
 
@@ -255,7 +314,7 @@ export class AppComponent implements OnInit {
   // fabmenu for mobile list
   onSelectMenu(event: any) {
     if (event.key === 'new') {
-      this.openCanvas(false);
+      this.openCanvas();
     }
   }
   // languageTableData: any = [] = [{ languagename: 'India', code: 'IND', isenabled: 'true', creationTime: '12-10-1992' },
