+++ raaghu-mfe/projects/language/src/app/app.module.ts
@@ -1,26 +1,19 @@
+import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
-import { BrowserModule } from '@angular/platform-browser';
 import { RdsButtonModule, RdsCheckboxModule, RdsDropdownlistModule, RdsDropdownModule, RdsIconModule, RdsOffcanvasModule, RdsSelectListModule } from '@libs/rds-elements';
 import { NgxTranslateModule, SharedModule } from '@libs/shared';
-import { LanguageEffects, LanguageReducer } from '@libs/state-management';
-import { EffectsModule } from '@ngrx/effects';
-import { StoreModule } from '@ngrx/store';
 import { RdsFabMenuModule } from 'projects/libs/rds-elements/src/rds-fab-menu/src/public-api';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { NewlanguageComponent } from './newlanguage/newlanguage.component';
-export const featureReducersMap = {
-  languages: LanguageReducer,
- 
-};
+
 @NgModule({
   declarations: [
     AppComponent,
-    NewlanguageComponent
   ],
   imports: [
     //BrowserModule,
+    CommonModule,
     AppRoutingModule,
     RdsOffcanvasModule,
     RdsButtonModule,
@@ -34,9 +27,6 @@ export const featureReducersMap = {
     RdsIconModule,
     NgxTranslateModule.forRoot(),
 
-    // StoreModule.forFeature('languages', featureReducersMap
-    // ),
-    // EffectsModule.forRoot([LanguageEffects]),
   ],
   // providers: [[HttpClient,
