--- raaghu-mfe/projects/language/src/app/newlanguage/newlanguage.component.html
@@ -1,1 +1,0 @@
-<mfe-loader [config]="rdsNewLanguageMfeConfig"></mfe-loader>
