+++ raaghu-mfe/projects/rds-components/src/app/rds-comp-new-language/rds-comp-new-language.component.ts
@@ -1,4 +1,4 @@
-import { Component, DoCheck, EventEmitter, Input, OnChanges, OnInit, Output, SimpleChanges } from '@angular/core';
+import { Component, EventEmitter, Input, OnChanges, OnInit, Output, SimpleChanges } from '@angular/core';
 import { DatePipe } from '@angular/common';
 import { NgForm } from '@angular/forms';
 import { TranslateService } from '@ngx-translate/core';
@@ -19,9 +19,11 @@ export class Language {
     DatePipe
   ]
 })
-export class RdsCompNewLanguageComponent implements OnInit, DoCheck, OnChanges {
+export class RdsCompNewLanguageComponent implements OnInit, OnChanges {
   constructor(public datepipe: DatePipe, public translate: TranslateService) { }
-  @Output() onLanguageSave = new EventEmitter<any>()
+  @Output() onLanguageSave = new EventEmitter<any>();
+  @Output() onCloseCanvas = new EventEmitter<any>()
+
   @Input() flags: any[] = []
   // @Input() selectedLanguage: any;
   @Input() languageNames: any[] = []
@@ -46,53 +48,15 @@ export class RdsCompNewLanguageComponent implements OnInit, DoCheck, OnChanges {
   }
 
   ngOnChanges(changes: SimpleChanges): void {
-    if (this.selectedLanguage.icon && this.flags.length > 0) {
-      const selectedLanguage = this.flags.find((x: any) => x.value === this.selectedLanguage.icon);
-      this.selectedLanguage.icon = selectedLanguage.some;
-    }
-    // if (this.selectedLanguage) {
-    //   this.name = this.selectedLanguage.countrycode;
-    //   this.icon = this.selectedLanguage.icon;
-    //   this.id = this.selectedLanguage.id;
-    //   this.isEnabled = (this.selectedLanguage.isDisabled) ? false : true;
-    // } else {
-    //   this.name = '';
-    //   this.icon = '';
-    //   this.id = undefined;
-    //   this.isEnabled = false;
-    // }
+
   }
 
 
-  ngDoCheck(): void {
-    if (this.selectedLanguage) {
-      this.name = this.selectedLanguage.countryCode;
-      this.icon = this.selectedLanguage.icon;
-      this.id = this.selectedLanguage.id;
-      this.isEnabled = (this.selectedLanguage.isEnabled) ? false : true;
-    } else {
-      this.name = '';
-      this.icon = '';
-      this.id = undefined;
-      this.isEnabled = false;
-    }
-  }
+
 
   ngOnInit(): void {
-    this.languageName = [];
-    this.languageName.push(`${this.selectedLanguage.countryCode}`);
-    if (this.selectedLanguage.icon && this.flags.length > 0) {
-      const selectedLanguage = this.flags.find((x: any) => x.value === this.selectedLanguage.icon);
-      this.selectedLanguage.icon = selectedLanguage.some;
-    }
     this.submitted = false;
-
-    if (this.selectedLanguage) {
-      this.name = this.selectedLanguage.countrycode;
-      this.icon = this.selectedLanguage.icon;
-      this.id = this.selectedLanguage.id;
-      this.isEnabled = (this.selectedLanguage.isDisabled) ? false : true;
-    }
+ 
   }
 
   addLanguage(languageForm: NgForm): void {
@@ -101,11 +65,15 @@ export class RdsCompNewLanguageComponent implements OnInit, DoCheck, OnChanges {
       return;
     }
     const flag: any = this.flags.find((x: any) => x.some === this.selectedLanguage.icon);
-    this.onLanguageSave.emit({ icon: flag.value, isEnabled: this.selectedLanguage.isEnabled, name: this.languageName[0], id: this.selectedLanguage.id });
+    this.onLanguageSave.emit({ icon: flag.value, isEnabled: this.selectedLanguage.isEnabled, name: this.selectedLanguage.countryCode[0], id: this.selectedLanguage.id });
 
   }
 
   onCountryCodeSelect(selectedItem: any): void {
     this.selectedLanguage.icon = selectedItem.item.some;
+  }
+
+  closeCanvas(): void {
+    this.onCloseCanvas.emit(true);
   }
 }
